refactor: Store check results and clean up code
# Why?
Checking the chip type and the kernel version over and over is slow and prone to mistakes along with the RMMIO/FB map attributes being incorrect and other logic issues.

# How?
Store the checks for use throughout the code, use the correct attributes to map calls and fix other logic issues.

Additional change: You may no longer set a custom model to prevent potential issues as it is not completely known whether this is purely cosmetic.